# server/src/main/resources/openapi/documentation.yaml
openapi: "3.0.3"
info:
  title: "Application API"
  description: "Application API"
  version: "1.0.0"
servers:
  - url: "http://127.0.0.1:8080"
paths:
...
  /github:
    get:
      tags:
        - GitHub Repository
      summary: Get GitHub Repositories with Pagination
      description: Retrieve a paginated list of GitHub repositories
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GHRepositoryMiniDao'
  /github/tags:
    get:
      tags:
        - GitHub Repository
      summary: Get Tags with Pagination
      description: Retrieve a paginated list of tags associated with GitHub repositories
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagDao'
components:
  schemas:
...
    TagDao:
      type: object
      properties:
        id:
          type: integer
          description: The id of the tag
        name:
          type: string
          description: The name of the tag
        description:
          type: string
          description: The description of the tag
        color:
          type: string
          description: The color of the tag
...
    GHRepositoryMiniDao:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier of the repository
        fullName:
          type: string
          description: The full name of the repository
        license:
          type: string
          nullable: true
          description: The license of the repository (optional)
        languages:
          type: object
          additionalProperties:
            type: integer
          description: A map of languages used in the repository and their line counts
        updatedAt:
          type: string
          format: date-time
          description: The date and time when the repository was last updated
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagDao'
          description: A list of tags associated with the repository